chore: organize pages
