+++ src/main.ts
@@ -1,11 +1,7 @@
 import "./app.css";
 import App from "./App.svelte";
-import Home from "./pages/Home.svelte";
+import Home from "./pages/HomePage.svelte";
 
 export const app = new App({
   target: document.getElementById("app"),
 });
-
-// export const home = new Home({
-//   target: document.getElementById('app')
-// })
