+++ src/pages/404Page.svelte
