+++ src/pages/FavoritesPage.svelte
